+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- _EqualityComparisonAssertion_ prohibits equality expression in `assertThat`
+
 ### Changed
 
 ### Deprecated